FeatureGenerator base class
